New check: Relying on the server time zone setting is unsafe.  Better to always explicitly code for a specific timezone.
Changes:

Emit Standard.Unsafe.TimeZone whenever encountering new DateTime($time), new DateTimeImmutable($time), or create_date($time), create_date_immutable($time) without a second parameter specifying the timezone.  There are actually a couple of scenarios where the time zone isn't necessary (when $time starts with the string "@" or ends with a time zone offset), but we can't verify the runtime value of a variable so we always emit the error.

Also, report an error anytime date() is called.  It always does a time zone conversion to the local time zone.