Update PSR-4 check to not use substring search
When a project uses a PSR-4 root name which is also the prefix of
another root name, the Guardrail check may choose the wrong root when
the shorter root appears first in the list. As a result, the PSR-4
check will resolve the wrong expected file path and emit a false
negative. I have updated the expected path resolution logic to mimic
Composer's own algorithm so that it is not affected by root name
combinations like this.

Jira: [GB-947]